==> 3.visualization.Rmd <==
---
title: "Obesity amongst men, NFHS - 5, visualisations"
author: "Arkaprabha Gun, Tushar Garg"
date: "`r Sys.Date()`"
---

```{r setup, include=F}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(echo=F,message=F, warning=F)
suppressMessages(if (!require("pacman")) install.packages("pacman"))
pacman::p_load(tidyverse,lubridate,rio,janitor,here,gtsummary,survey,haven,labelled,sjlabelled,performance,jtools,forestploter,svglite)
```

```{r import}
load(here("Output","men_output_viz.Rdata"));load(here("Output","women_output_viz.Rdata"))
```

```{r original paper data}
# Importing the prevalnce and model estiamtes tables from the originsal paper by Chaudhary et al
orig_prev <- import(here("Data_raw","original paper data_chaudhary.xlsx"), which = "prevalence")
orig_model <- import(here("Data_raw","original paper data_chaudhary.xlsx"), which = "model")

# Cleaning and aligning the table for the forest plot
orig_prev <- orig_prev %>%
  extract("Women-WC", into = c("orig_prop_Obese_wc_women", "orig_low_ci_Obese_wc_women", "orig_high_ci_Obese_wc_women"),
          regex = "([0-9.]+)\\(([0-9.]+)\\–([0-9.]+)\\)", remove = FALSE) %>%
  extract("Women-BMI", into = c("orig_prop_Obese_bmi_women", "orig_low_ci_Obese_bmi_women", "orig_high_ci_Obese_bmi_women"),
          regex = "([0-9.]+)\\(([0-9.]+)\\–([0-9.]+)\\)", remove = FALSE) %>%
  extract("Men-WC", into = c("orig_prop_Obese_wc_men", "orig_low_ci_Obese_wc_men", "orig_high_ci_Obese_wc_men"),
          regex = "([0-9.]+)\\(([0-9.]+)\\–([0-9.]+)\\)", remove = FALSE) %>%
  extract("Men-BMI", into = c("orig_prop_Obese_bmi_men", "orig_low_ci_Obese_bmi_men", "orig_high_ci_Obese_bmi_men"),
          regex = "([0-9.]+)\\(([0-9.]+)\\–([0-9.]+)\\)", remove = FALSE)
```

```{r prevalence}
process_prev <- function(data, data_wc, data_bmi, gender) {
  
  fig_prev <- data %>% 
    as_data_frame() %>% 
    rename("Characteristic"=1, "count_prop"=2, "ci"=3) %>% 
    separate(count_prop, into = c("count", "prop"), sep = " \\("
    )%>% 
    mutate(
      prop = str_extract(prop, "\\d+\\·?\\d*"),
      prop = str_replace_all(prop, "\\·", "."),
      prop = as.numeric(prop)/100,
      count = as.numeric(str_replace_all(count,",", ""))
    ) %>%
    separate(ci, into = c("low_ci", "high_ci"), sep = ", ") %>%
    mutate(
      low_ci = as.numeric(str_replace(str_remove(low_ci, "%"), "\\·", ".")) / 100,
      high_ci = as.numeric(str_replace(str_remove(high_ci, "%"), "\\·", ".")) / 100
    ) %>%
    mutate(Characteristic = case_when(
      row_number(Characteristic) %in% 1:3 ~ paste0(Characteristic, "_wc_", gender),
      row_number(Characteristic) %in% 4:6 ~ paste0(Characteristic, "_bmi_", gender),
      TRUE ~ Characteristic
    )) %>%
    filter(str_detect(Characteristic, "Obese")) %>%
    pivot_wider(
      names_from = Characteristic,
      values_from = c(count, prop, low_ci, high_ci)
    ) %>%
    mutate(Characteristic = "Overall") %>%
    select(Characteristic, everything())
  
  # Code for wc
  fig_prev_wc <- data_wc %>%
    as_data_frame() %>%
    select(-contains(c("Non-","Overall"))) %>%
    rename("Characteristic"=1, "estimate"=2) %>%
    separate(estimate, into = c("prop", "count", "low_ci", "high_ci"), sep = " ") %>% 
    mutate(prop = str_extract(prop, "\\d+\\·?\\d*") %>% as.numeric() / 100,
           count = str_extract(count, "\\d{1,3}(?:,\\d{3})*") %>% 
             gsub(",", "", .) %>% 
             as.numeric(),
           low_ci = str_extract(low_ci, "(?<=\\[)\\d+") %>% as.numeric() / 100,
           high_ci = str_extract(high_ci, "\\d+(?=\\%\\])") %>% as.numeric() / 100) %>% 
    rename_with(~paste0(., paste0("_Obese_wc_", gender)), .cols = -c("Characteristic"))
  
  # Code for bmi
  fig_prev_bmi <- data_bmi %>%
    as_data_frame() %>%
    select(-contains(c("Non-","Overall"))) %>%
    rename("Characteristic"=1, "estimate"=2) %>%
    separate(estimate, into = c("prop", "count", "low_ci", "high_ci"), sep = " ") %>% 
    mutate(prop = str_extract(prop, "\\d+\\·?\\d*") %>% as.numeric() / 100,
           count = str_extract(count, "\\d{1,3}(?:,\\d{3})*") %>% 
             gsub(",", "", .) %>% 
             as.numeric(),
           low_ci = str_extract(low_ci, "(?<=\\[)\\d+") %>% as.numeric() / 100,
           high_ci = str_extract(high_ci, "\\d+(?=\\%\\])") %>% as.numeric() / 100) %>%
    rename_with(~paste0(., paste0("_Obese_bmi_", gender)), .cols = -c("Characteristic"))
  
  # Final assembly
  fig_prev <- left_join(fig_prev_wc, fig_prev_bmi, by = "Characteristic") %>%
    bind_rows(fig_prev, .) %>%
    select(-paste0("count_Obese_bmi_", gender))
  
  return(fig_prev)
}

fig_prev_women <- process_prev(women_prev, women_prev_wc, women_prev_bmi, "women")
fig_prev_men <- process_prev(men_prev, men_prev_wc, men_prev_bmi, "men")

fig_prev <- left_join(fig_prev_women, fig_prev_men, by = "Characteristic") %>% 
  filter(`Characteristic` != "NA")

```

```{r prevalence figure}
# indent the subgroup
indent_check <- c("Overall","Age category", "Type of Residence", "Highest Education Level", 
                  "Religion", "Wealth Index", "Caste", "Frequency of Eating Meat")

fig_prev$Characteristic <- ifelse(fig_prev$Characteristic %in% indent_check, fig_prev$Characteristic,
                                  paste0("     ", fig_prev$Characteristic))

# NA to blank or NA will be transformed to character.
fig_prev$"n[obese-wc]" <- ifelse(is.na(fig_prev$count_Obese_wc_women), "", fig_prev$count_Obese_wc_women)
fig_prev$"n[obese-wc ]" <- ifelse(is.na(fig_prev$count_Obese_wc_men), "", fig_prev$count_Obese_wc_men)


# Add two blank column for CI
fig_prev$`Prevalence` <- paste(rep(" ", 20), collapse = " ")
fig_prev$`Prevalence ` <- paste(rep(" ", 20), collapse = " ")

# Set-up theme
theme_prev <- forest_theme(base_size = 13,
                           base_family = "Helevetica",
                           refline_lwd = 0,
                           refline_lty = "solid",
                           ci_pch = c(2, 1),
                           ci_lty = c("solid","solid"),
                           ci_col = c("#8D0522","#05668D"),
                           ci_lwd = c(1.3,0.85),
                           ci_Theight = NULL,
                           footnote_col = "blue",
                           legend_name = "Criterion",
                           legend_value = c(" Waist circumference   ", " Body mass index"),
                           legend_position = "bottom",
                           vertline_lty = c("dashed"),
                           vertline_lwd = 1,
                           vertline_col = c("#7ABA69"),
                           colhead=list(fg_params=list(
                             hjust=c(0.83, 0.5, 0.5, 0.5, 0.5),
                             x=0.5, parse=T)))

fig_prev_plot <- forest(fig_prev[,c(1, 16, 18, 17, 19)],
                        est = list(fig_prev$prop_Obese_wc_women,
                                   fig_prev$prop_Obese_wc_men,
                                   fig_prev$prop_Obese_bmi_women,
                                   fig_prev$prop_Obese_bmi_men),
                        lower = list(fig_prev$low_ci_Obese_wc_women,
                                     fig_prev$low_ci_Obese_wc_men,
                                     fig_prev$low_ci_Obese_bmi_women,
                                     fig_prev$low_ci_Obese_bmi_men), 
                        upper = list(fig_prev$high_ci_Obese_wc_women,
                                     fig_prev$high_ci_Obese_wc_men,
                                     fig_prev$high_ci_Obese_bmi_women,
                                     fig_prev$high_ci_Obese_bmi_men), 
                        ci_column = c(3, 5),
                        xlim = list(c(0, 1), c(0, 1)),
                        xlab = c(""),
                        ticks_at = list(c(0, 0.25, 0.5, 0.75, 1),c(0, 0.25, 0.5, 0.75, 1)),
                        vert_line = list(c(fig_prev$prop_Obese_wc_women[fig_prev$Characteristic == "Overall"]), 
                                         c(fig_prev$prop_Obese_wc_men[fig_prev$Characteristic == "Overall"])),
                        nudge_y = 0.25,
                        theme = theme_prev) %>% 
  # bold category heading
  edit_plot(row = c(1,2,6,9,14,22,28,34),
            gp = gpar(fontface = "bold")) %>% 
  # right aligned numbers
  edit_plot(col = c(2,3,4,5), which = "text",
            hjust = unit(1, "npc"), 
            x = unit(0.9, "npc")) %>% 
  # all background white
  edit_plot(row = c(1:42), which = "background",
            gp = gpar(fill = "white")) %>% 
  # background to category headings
  edit_plot(row = c(1,2,6,9,14,22,28,34), which = "background",
            gp = gpar(fill = "lightgray")) %>% 
  # background to overall row
  edit_plot(col = 1:5, row = 1, which = "background", 
            gp = gpar(fill = "#f6eff7")) %>%
  # header join: women and men
  insert_text(text = "Women",col = 2:3,part = "header",
              gp = gpar(fontface = "bold")) %>%
  add_text(text = "Men",col = 4:5,part = "header",row = 1,
           gp = gpar(fontface = "bold")) %>%
  edit_plot(part = "header", row = 2,
            gp = gpar(fontface = "bold")) %>% 
  
  # top border for table
  add_border(part = "header", row = 1, where = "top",
             gp = gpar(lwd = .7)) %>% 
  # bottom border for header
  add_border(part = "header", row = 2, where = "bottom",
             gp = gpar(lwd = .7)) %>% 
  # border for women header
  add_border(part = "header", row = 1, col = 2:3,
             gp = gpar(lwd = 0.8)) %>% 
  # border for men header
  add_border(part = "header", row = 1, col = 4:5,
             gp = gpar(lwd = 0.8))

ggplot2::ggsave(filename = here("Output", "fig_prev_plot.svg"), plot = fig_prev_plot,
                dpi = 300, width = 8.3, height = 13.6, units = "in")
```

```{r prevalence figure with original estimates}
orig_prev$Characteristic <- ifelse(orig_prev$Characteristic %in% indent_check, orig_prev$Characteristic,
                                  paste0("     ", orig_prev$Characteristic))

fig_prev2 <- fig_prev %>%
  left_join(orig_prev %>% select(Characteristic, starts_with("orig_")), by = "Characteristic") %>% 
  mutate(across(starts_with("orig_"), ~ as.numeric(.) / 100)) %>% 
  select(-(starts_with("orig_") & contains("bmi")))

theme_prev1 <- forest_theme(base_size = 13,
                            base_family = "Helevetica",
                            refline_lwd = 0,
                            refline_lty = "solid",
                            ci_pch = c(2, 1, 2),
                            ci_lty = c("solid","solid","solid"),
                            ci_col = c("#8D0522","#034762","#AF5064"),
                            ci_lwd = c(1.3,0.85,0.65),
                            ci_Theight = NULL,
                            footnote_col = "blue",
                            legend_name = "Criterion",
                            legend_value = c(" Waist circumference   ", " Body mass index", ""),
                            legend_position = "bottom",
                            vertline_lty = c("dashed"),
                            vertline_lwd = 1,
                            vertline_col = c("#7ABA69"),
                            colhead=list(fg_params=list(
                              hjust=c(0.83, 0.5, 0.5, 0.5, 0.5),
                              x=0.5, parse=T)),
                            core = list(padding = unit(c(0,5), "mm")) 
                            )

fig_prev_plot2 <- forest(
  fig_prev2[, c(1, 16, 18, 17, 19)],
  est = list(
    fig_prev2$prop_Obese_wc_women,
    fig_prev2$prop_Obese_wc_men,
    fig_prev2$prop_Obese_bmi_women,
    fig_prev2$prop_Obese_bmi_men,
    fig_prev2$orig_prop_Obese_wc_women,  # original estimates
    fig_prev2$orig_prop_Obese_wc_men
    # ,
    # fig_prev2$orig_prop_Obese_bmi_women,
    # fig_prev2$orig_prop_Obese_bmi_men
  ),
  lower = list(
    fig_prev2$low_ci_Obese_wc_women,
    fig_prev2$low_ci_Obese_wc_men,
    fig_prev2$low_ci_Obese_bmi_women,
    fig_prev2$low_ci_Obese_bmi_men,
    fig_prev2$orig_low_ci_Obese_wc_women,  # original confidence intervals
    fig_prev2$orig_low_ci_Obese_wc_men
    # ,
    # fig_prev2$orig_low_ci_Obese_bmi_women,
    # fig_prev2$orig_low_ci_Obese_bmi_men
  ), 
  upper = list(
    fig_prev2$high_ci_Obese_wc_women,
    fig_prev2$high_ci_Obese_wc_men,
    fig_prev2$high_ci_Obese_bmi_women,
    fig_prev2$high_ci_Obese_bmi_men,
    fig_prev2$orig_high_ci_Obese_wc_women,  # original confidence intervals
    fig_prev2$orig_high_ci_Obese_wc_men
    # ,
    # fig_prev2$orig_high_ci_Obese_bmi_women,
    # fig_prev2$orig_high_ci_Obese_bmi_men
  ), 
  ci_column = c(3, 5),
  xlim = list(c(0, 1), c(0, 1)),
  xlab = c(""),
  ticks_at = list(c(0, 0.25, 0.5, 0.75, 1),c(0, 0.25, 0.5, 0.75, 1)),
  vert_line = list(c(fig_prev2$prop_Obese_wc_women[fig_prev2$Characteristic == "Overall"]), 
                   c(fig_prev2$prop_Obese_wc_men[fig_prev2$Characteristic == "Overall"])),
  nudge_y = 0.25,
  theme = theme_prev1
) %>%  # bold category heading
  edit_plot(row = c(1,2,6,9,14,22,28,34),
            gp = gpar(fontface = "bold")) %>%
  # right aligned numbers
  edit_plot(col = c(2,3,4,5), which = "text",
            hjust = unit(1, "npc"), 
            x = unit(0.9, "npc")) %>% 
  # all background white
  edit_plot(row = c(1:42), which = "background",
            gp = gpar(fill = "white")) %>% 
  # background to category headings
  edit_plot(row = c(1,2,6,9,14,22,28,34), which = "background",
            gp = gpar(fill = "lightgray")) %>% 
  # background to overall row
  edit_plot(col = 1:5, row = 1, which = "background", 
            gp = gpar(fill = "#f6eff7")) %>%
  # header join: women and men
  insert_text(text = "Women",col = 2:3,part = "header",
              gp = gpar(fontface = "bold")) %>%
  add_text(text = "Men",col = 4:5,part = "header",row = 1,
           gp = gpar(fontface = "bold")) %>%
  edit_plot(part = "header", row = 2,
            gp = gpar(fontface = "bold")) %>% 
  
  # top border for table
  add_border(part = "header", row = 1, where = "top",
             gp = gpar(lwd = .7)) %>% 
  # bottom border for header
  add_border(part = "header", row = 2, where = "bottom",
             gp = gpar(lwd = .7)) %>% 
  # border for women header
  add_border(part = "header", row = 1, col = 2:3,
             gp = gpar(lwd = 0.8)) %>% 
  # border for men header
  add_border(part = "header", row = 1, col = 4:5,
             gp = gpar(lwd = 0.8))

ggplot2::ggsave(filename = here("Output", "fig_prev_plot2.svg"), plot = fig_prev_plot2,
                dpi = 300, width = 8.3, height = 14, units = "in")
```


```{r model}
process_model <- function(data, gender_suffix, model_suffix) {
  data %>%
    as_data_frame() %>% 
    rename(Characteristic = `**Characteristic**`,
           or = `**OR**`,
           ci = `**95% CI**`,
           p_value = `**p-value**`) %>%
    separate(ci, into = c("low_ci", "high_ci"), sep = ", ") %>%
    mutate(
      !!paste0("or_", gender_suffix, "_", model_suffix) := str_replace_all(or, "\\·", ".") %>% as.numeric(),
      !!paste0("low_ci_", gender_suffix, "_", model_suffix) := str_replace_all(low_ci, "\\·", ".") %>% as.numeric(),
      !!paste0("high_ci_", gender_suffix, "_", model_suffix) := str_replace_all(high_ci, "\\·", ".") %>% as.numeric(),
      !!paste0("p_value_", gender_suffix, "_", model_suffix) := as.numeric(p_value)
    ) %>%
    select(Characteristic, 
           paste0("or_", gender_suffix, "_", model_suffix), 
           paste0("low_ci_", gender_suffix, "_", model_suffix), 
           paste0("high_ci_", gender_suffix, "_", model_suffix), 
           paste0("p_value_", gender_suffix, "_", model_suffix))
}

fig_model_women_adj <- process_model(women_model_adj_wc_est, "women", "adj")
fig_model_men_adj <- process_model(men_model_adj_wc_est, "men", "adj")

fig_model_women_uv <- process_model(women_model_uv_wc, "women", "uv")
fig_model_men_uv <- process_model(men_model_uv_wc, "men", "uv")

fig_model <- left_join(fig_model_women_adj, fig_model_women_uv, by = "Characteristic") %>%
  left_join(fig_model_men_adj, by = "Characteristic") %>%
  left_join(fig_model_men_uv, by = "Characteristic") %>% 
  select(-contains("p_value"))
```

```{r model figure}
# indent the subgroup
indent_check <- c("Age category", "Type of Residence", "Highest Education Level", 
                  "Religion", "Wealth Index", "Caste", "Frequency of Eating Meat")

fig_model$Characteristic <- ifelse(fig_model$Characteristic %in% indent_check, fig_model$Characteristic,
                                   paste0("     ", fig_model$Characteristic))


fig_model$`adj OR (95% CI)` <- ifelse(is.na(fig_model$or_women_adj), "",
                                      sprintf(" %.2f (%.2f to %.2f)  ",
                                              fig_model$or_women_adj,
                                              fig_model$low_ci_women_adj, fig_model$high_ci_women_adj))

fig_model$`adj OR (95% CI) ` <- ifelse(is.na(fig_model$or_men_adj), "",
                                       sprintf(" %.2f (%.2f to %.2f)",
                                               fig_model$or_men_adj,
                                               fig_model$low_ci_men_adj, fig_model$high_ci_men_adj))

# adding text: Reference
fig_model <- fig_model %>%
  mutate(`adj OR (95% CI)` = ifelse(!Characteristic %in% indent_check & is.na(or_women_adj), " Reference", `adj OR (95% CI)`), 
         `adj OR (95% CI) ` = ifelse(!Characteristic %in% indent_check & is.na(or_women_adj), " Reference", `adj OR (95% CI) `))

# Add two blank column for CI
fig_model$` ` <- paste(rep(" ", 32), collapse = " ") #Women
fig_model$`  ` <- paste(rep(" ", 32), collapse = " ") #Men

# Set-up theme
theme_model <- forest_theme(base_size = 13,
                            base_family = "Helevetica",
                            refline_lwd = 0,
                            refline_lty = "solid",
                            ci_pch = c(5,19),
                            ci_lty = c("solid","solid"),
                            ci_col = c("#D81B60", "#0095A8"),
                            ci_alpha = c(0.9),
                            ci_lwd = c(0.6, 1.3),
                            ci_Theight = NULL,
                            footnote_col = "blue",
                            legend_name = "Model",
                            legend_value = c(" Univariate  ", "  Adjusted"),
                            legend_position = "bottom",
                            vertline_lty = c("dashed"),
                            vertline_lwd = 1.2,
                            vertline_col = c("#EEB819"),
                            colhead=list(fg_params=list(
                              hjust=c(0.83, 0.5, 0.5, 0.5, 0.5),
                              x=0.5, parse=T))
                            )

fig_model_plot <- forest(fig_model[,c(1, 14, 16, 15, 17)],
                         est = list(fig_model$or_women_uv,
                                    fig_model$or_men_uv,
                                    fig_model$or_women_adj,
                                    fig_model$or_men_adj),
                         lower = list(fig_model$low_ci_women_uv,
                                      fig_model$low_ci_men_uv,
                                      fig_model$low_ci_women_adj,
                                      fig_model$low_ci_men_adj), 
                         upper = list(fig_model$high_ci_women_uv,
                                      fig_model$high_ci_men_uv,
                                      fig_model$high_ci_women_adj,
                                      fig_model$high_ci_men_adj), 
                         ci_column = c(3, 5),
                         xlim = list(c(0, 6), c(0, 6)),
                         xlab = c(""),
                         ticks_at = list(c(0, 1, 2, 3, 4, 5, 6),c(0, 1, 2, 3, 4, 5, 6)),
                         vert_line = list(c(NA,1), c(NA,1)),
                         nudge_y = 0.25,
                         theme = theme_model) %>%
  # bold category heading
  edit_plot(row = c(1,5,8,13,21,27,33),
            gp = gpar(fontface = "bold")) %>% 
  # left aligned numbers
  edit_plot(col = c(2,3,4,5), which = "text",
            hjust = unit(0, "npc"),
            x = unit(0, "npc")) %>% 
  # all background white
  edit_plot(row = c(1:40), which = "background",
            gp = gpar(fill = "white")) %>%
  # background to category headings
  edit_plot(row = c(1,5,8,13,21,27,33), which = "background",
            gp = gpar(fill = "lightgray")) %>% 
  # header join: women and men
  insert_text(text = "Women",col = 2:3,part = "header",
              gp = gpar(fontface = "bold")) %>%
  add_text(text = "Men",col = 4:5,part = "header",row = 1,
           gp = gpar(fontface = "bold")) %>%
  edit_plot(part = "header", row = 2,
            gp = gpar(fontface = "bold")) %>% 
  # top border for table
  add_border(part = "header", row = 1, where = "top",
             gp = gpar(lwd = .7)) %>%
  # bottom border for header
  add_border(part = "header", row = 2, where = "bottom",
             gp = gpar(lwd = .7)) %>%
  # border for women header
  add_border(part = "header", row = 1, col = 2:3,
             gp = gpar(lwd = 0.8)) %>%
  # border for men header
  add_border(part = "header", row = 1, col = 4:5,
             gp = gpar(lwd = 0.8))

ggplot2::ggsave(filename = here("Output", "fig_model_plot.svg"), plot = fig_model_plot,
                dpi = 300, width = 13, height = 13, units = "in")

```

```{r missing caste values}
process_df <- function(df, name) {
  # Filter rows where "**Characteristic**" contains the string "NA"
  df <- df[df$`**Characteristic**` == "NA", ]
  
  # Clean up column names
  colnames(df) <- sapply(colnames(df), function(x) {
    if (grepl("Non-obese", x)) "Non-obese"
    else if (grepl("Obese", x)) "Obese"
    else if (grepl("Overall", x)) "Total"
    else x
  })
  
  # Clean df_name column based on provided conditions
  cleaned_name <- switch(name,
    "women_prev_wc" = "Women Waist",
    "women_prev_bmi" = "Women BMI",
    "men_prev_wc" = "Men Waist",
    "men_prev_bmi" = "Men BMI",
    name # default case
  )
  
  df$missing <- cleaned_name
  
  return(df)
}

# Process & bind all data frames together
caste_missing_values <- bind_rows(process_df(women_prev_wc, "women_prev_wc"),
                           process_df(women_prev_bmi, "women_prev_bmi"),
                           process_df(men_prev_wc, "men_prev_wc"),
                           process_df(men_prev_bmi, "men_prev_bmi")) %>% 
  select(-`**Characteristic**`)


# Helper function to convert character values to numeric
convert_to_numeric <- function(value) {
  as.numeric(gsub(",", "", value))
}

# Adjusting the extraction function
extract_values <- function(category, metric) {
  value <- caste_missing_values %>%
    filter(missing == category) %>%
    pull(!!as.name(metric))
  convert_to_numeric(value)
}

# Women Waist
women_waist_total <- extract_values("Women Waist", "Total")
women_waist_obese <- extract_values("Women Waist", "Obese")

# Women BMI
women_bmi_total <- extract_values("Women BMI", "Total")
women_bmi_obese <- extract_values("Women BMI", "Obese")

# Men Waist
men_waist_total <- extract_values("Men Waist", "Total")
men_waist_obese <- extract_values("Men Waist", "Obese")

# Men BMI
men_bmi_total <- extract_values("Men BMI", "Total")
men_bmi_obese <- extract_values("Men BMI", "Obese")

sprintf(
  "Caste data was missing for %s women and %s men, respectively.",
  formatC(women_waist_total, format="d", big.mark=','),
  formatC(men_waist_total, format="d", big.mark=',')
)
sprintf(
  "Caste data was missing for %s (%.2f%%) and %s (%.2f%%) women classified as obese based on waist and BMI, respectively. For men, caste was missing for %s (%.2f%%) and %s (%.2f%%) individuals classified as obese based on waist and BMI, respectively.", 
  formatC(women_waist_obese, format="d", big.mark=','), 
  (women_waist_obese/women_waist_total)*100, 
  formatC(women_bmi_obese, format="d", big.mark=','), 
  (women_bmi_obese/women_bmi_total)*100,
  formatC(men_waist_obese, format="d", big.mark=','), 
  (men_waist_obese/men_waist_total)*100, 
  formatC(men_bmi_obese, format="d", big.mark=','), 
  (men_bmi_obese/men_bmi_total)*100)
```


```{r session info}
sessioninfo::session_info(pkgs = c("loaded", "attached")[1], to_file = here("Session info","session_info_visualization.txt"))
```
