==> 4.maps.Rmd <==
---
title: "Obesity amongst men, NFHS - 5, generate maps"
author: "Arkaprabha Gun, Tushar Garg"
date: "`r Sys.Date()`"
---

```{r setup, include=F}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_chunk$set(echo=F,message=F, warning=F)
suppressMessages(if (!require("pacman")) install.packages("pacman"))
p_load(tidyverse, rio, janitor, here, geojsonio, ggmap, ggthemes, maptiles, MetBrewer, sf, tidyterra, svglite, sessioninfo)
```

```{r loading  shapefile, include=FALSE}
# DHS shapefile for India 2020 DHS, available at https://spatialdata.dhsprogram.com/boundaries/#view=table&countryId=IA
# Unzip in the Data_raw folder
district_shapefile <- sf::st_read(here("Data_raw","sdr_subnational_boundaries_2023-08-21","shps","sdr_subnational_boundaries2.shp"))

# background map of India with correct geography
underlay_map <- st_read(here("Data_raw","shapefile","india-composite.geojson"))
# map with state boundaries
states <- geojson_sf("Data_raw/shapefile/geoBoundaries-IND-ADM1_simplified.geojson")
```

```{r importing obesity estimates}
state_men <- import(here("Output","men_state_obesity_prev_wc.xlsx"))
state_women <- import(here("Output","women_state_obesity_prev_wc.xlsx"))

district_men <- import(here("Output","men_dist_obesity_prev_wc.xlsx"))
district_women <- import(here("Output","women_dist_obesity_prev_wc.xlsx"))
```

```{r color palette}
# met.brewer(name="Demuth", n=7, type="discrete", direction = -1)
color_palette <- c("<10" = "#41485f", "10-20" = "#5d6174", "20-30" = "#6e8291", "30-40" = "#f7c267", 
                   "40-50" = "#d39a2d", "50-60" = "#b64f32", ">60" = "#9b332b")
```

############
#District
############

```{r merging men and womens district level estimates}
district_merged_data <- merge(district_men, district_women, by = c("district_id","state","district"), suffixes = c("_men", "_women"))
```

```{r cleaning district names to match with DHS shapefile}
district_merged_data <- district_merged_data %>%
  mutate(
    district = case_when(
      district == "North  & Middle Andaman" ~ "North & Middle Andaman",
      district == "Janjgir - Champa" ~ "Janjgir-Champa",
      district == "Buxar" ~ "Buxer",
      district == "Leh(Ladakh)" ~ "Leh",
      district == "North  District" ~ "North District",
      district == "Mahrajganj" ~ "Maharajganj",
      district == "Sant Ravidas Nagar (Bhadohi)" ~ "Sant Ravidas Nagar",
      district == "Y.s.r." ~ "Y.S.R.",
      TRUE ~ district  # Keep the original value for all other cases
    ),
    state = case_when(
      state == "Dadra & Nagar Haveli And Daman & Diu" ~ "Dadra & Nagar Haveli & Daman & Diu",
      TRUE ~ state  # Keep the original value for all other cases
    )
  )

district_merged_data <- district_merged_data %>%
  mutate(cat_obese_women = cut(prev_obese_women, 
                               breaks = c(-Inf, 10, 20, 30, 40, 50, 60, Inf), 
                               labels = c("<10", "10-20", "20-30", "30-40", "40-50", "50-60", ">60"),
                               right = FALSE)
  )
```

```{r renaming district and state name in dhs shapefile}
district_shapefile <- district_shapefile %>%
  rename(district = REGNAME, state = OTHREGNA)

# merging district shapefile with district-level estimates
district_merged_data_with_shape <- merge(district_merged_data, district_shapefile, by = c("district","state"))
```

```{r manipulating shapefiles for concurrence}
map_extent <- as.numeric(st_bbox(underlay_map))
grid_box = c(map_extent[1], map_extent[2], map_extent[3], map_extent[4])
class(grid_box) <- "bbox"
grid_poly <- st_as_sfc(grid_box) %>% st_set_crs(4326) %>%
  st_transform(3857)
# basemap <- get_tiles(grid_poly, provider = "Stamen.Watercolor", crop = TRUE)
```

```{r converting to sf object and plotting mens estimates choropleth}
# While we generate the district-level map for men, these estimates should be used with caution because of small sample size for men at the districtplevel.

# Create an sf object with valid geometry
district_merged_data_with_shape_sf <- st_as_sf(district_merged_data_with_shape)

# Create the static map using ggplot2
district_menplot <- ggplot() + 
  geom_sf(data = underlay_map,aes(color ="No Data",pattern = "stripe")) +
  scale_color_manual("",values = c("No Data" = "black"),
                     guide = guide_legend(override.aes = list(linetype = "solid", shape = NA))) +  
  geom_sf(data = district_merged_data_with_shape_sf, aes(fill = prev_obese_men)) +
  scale_fill_viridis_c(direction = -1, limits = c(0,78.5))+
  theme_void()+  
  guides(fill = guide_colorbar(title.position = "top", title.hjust = 0.5)) +
  labs(title = "Prevalence of Abdominal Obesity in Men by District",fill = "Prevalence in % age")

district_menplot

# ggsave(here("Output","map_district_men.svg"), plot = district_menplot, device = "svg", width = 10, height = 8, units = "in")
```

```{r plotting womens district level estimates}
# Create the static map using ggplot2
district_womenplot <- ggplot() + 
  geom_sf(data = underlay_map,aes(color ="No Data")) +
  scale_color_manual("",values = c("No Data" = "black"),
                     guide = guide_legend(override.aes = list(linetype = "solid", shape = NA))) +
  geom_sf(data = district_merged_data_with_shape_sf, aes(fill = prev_obese_women)) +
  scale_fill_viridis_c( direction = -1, limits = c(0,78.5))+
  theme_void()+  
  guides(fill = guide_colorbar(title.position = "top", title.hjust = 0.5)) +
  labs(title = "Prevalence of Abdominal Obesity in women by District",fill = "Prevalence in % age")

district_womenplot

ggsave(here("Output","map_district_women.svg"), plot = district_womenplot, device = "svg", width = 10, height = 8, units = "in")

# Alternate map: abdominal obesity category
district_womenplot_cat <- ggplot() + 
  geom_sf(data = underlay_map,aes(color ="No Data")) +
  scale_color_manual("",values = c("No Data" = "black"),
                     guide = guide_legend(override.aes = list(linetype = "solid", shape = NA))) +
  geom_sf(data = district_merged_data_with_shape_sf, aes(fill = cat_obese_women)) +
  scale_fill_manual(values = color_palette) +
  theme_void()+  
  labs(title = "Prevalence of Abdominal Obesity in women by District",fill = "Prevalence in % age")

district_womenplot_cat

ggsave(here("Output","map_district_women_cat.svg"), plot = district_womenplot_cat, device = "svg", width = 10, height = 8, units = "in")
```

############
#State
############

```{r merging men and womens state level estimates}
state_merged_data_state <- merge(state_men, state_women, by = c("state"), suffixes = c("_men", "_women"))

# correcting names in shape file
states_corrected <- states %>%
  mutate(
    shapeName = case_when(
      shapeName == "Mahārāshtra" ~ "Maharashtra",
      shapeName == "Tamil Nādu" ~ "Tamil nadu",
      shapeName == "Chhattīsgarh" ~ "Chhattisgarh",
      shapeName == "Gujarāt" ~ "Gujarat",
      shapeName == "Dādra and Nagar Haveli and Damān and Diu" ~ "Dadra & nagar haveli and daman & diu",
      shapeName == "Himāchal Pradesh" ~ "Himachal pradesh",
      shapeName == "Jammu and Kashmīr" ~ "Jammu & kashmir",
      shapeName == "Bihār" ~ "Bihar",
      shapeName == "Ladākh" ~ "Ladakh",
      shapeName == "Rājasthān" ~ "Rajasthan",
      shapeName == "Telangāna" ~ "Telangana",
      shapeName == "Andhra Pradesh" ~ "Andhra pradesh",
      shapeName == "Arunāchal Pradesh" ~ "Arunachal pradesh",
      shapeName == "Meghālaya" ~ "Meghalaya",
      shapeName == "Jhārkhand" ~ "Jharkhand",
      shapeName == "Manipur" ~ "Manipur",
      shapeName == "Karnātaka" ~ "Karnataka",
      shapeName == "Uttarākhand" ~ "Uttarakhand",
      shapeName == "Uttar Pradesh" ~ "Uttar pradesh",
      shapeName == "Nāgāland" ~ "Nagaland",
      shapeName == "Lakshadweep" ~ "Lakshadweep",
      shapeName == "Sikkim" ~ "Sikkim",
      shapeName == "Tripura" ~ "Tripura",
      shapeName == "West Bengal" ~ "West bengal",
      shapeName == "Odisha" ~ "Odisha",
      shapeName == "Andaman and Nicobar Islands" ~ "Andaman & nicobar islands",
      shapeName == "Puducherry" ~ "Puducherry",
      shapeName == "Punjab" ~ "Punjab",
      shapeName == "Haryāna" ~ "Haryana",
      shapeName == "Delhi" ~ "Nct of delhi",
      shapeName == "Madhya Pradesh" ~ "Madhya pradesh",
      shapeName == "Chandīgarh" ~ "Chandigarh",
      
      # Add more corrections as needed
      TRUE ~ shapeName
    )
  )

state_merged_data_with_state <- merge(state_merged_data_state, states_corrected, by.x = "state", by.y = "shapeName")

state_merged_data_with_state <- state_merged_data_with_state %>%
  mutate(cat_obese_men = cut(prev_obese_men, 
                             breaks = c(-Inf, 10, 20, 30, 40, 50, 60, Inf), 
                             labels = c("<10", "10-20", "20-30", "30-40", "40-50", "50-60", ">60"),
                             right = FALSE),
         cat_obese_women = cut(prev_obese_women, 
                               breaks = c(-Inf, 10, 20, 30, 40, 50, 60, Inf), 
                               labels = c("<10", "10-20", "20-30", "30-40", "40-50", "50-60", ">60"),
                               right = FALSE)
  )

state_merged_data_with_shape_state_sf <- st_as_sf(state_merged_data_with_state)
```

```{r state plot for men}
state_menplot <- ggplot() + 
  geom_sf(data = state_merged_data_with_shape_state_sf, aes(fill = prev_obese_men)) +
  scale_fill_viridis_c( direction = -1, limits = c(0,78.5))+
  theme_void()+  
  guides(fill = guide_colorbar(title.position = "top", title.hjust = 0.5)) +
  labs(title = "Prevalence of Abdominal Obesity in Men by state",fill = "Prevalence in % age")

state_menplot

ggsave(here("Output","map_state_men.svg"), plot = state_menplot, device = "svg", width = 10, height = 8, units = "in")

# Alternate map: abdominal obesity category
state_menplot_cat <- ggplot() + 
  geom_sf(data = state_merged_data_with_shape_state_sf, aes(fill = cat_obese_men)) +
  scale_fill_manual(values = color_palette) +
  theme_void() +  
  labs(title = "Prevalence of Abdominal Obesity in Men by state", fill = "Prevalence in % age")

state_menplot_cat

ggsave(here("Output","map_state_men_cat.svg"), plot = state_menplot_cat, device = "svg", width = 10, height = 8, units = "in")
```

```{r state plot for women}
state_womenplot <- ggplot() + 
  geom_sf(data = state_merged_data_with_shape_state_sf, aes(fill = prev_obese_women)) +
  scale_fill_viridis_c( direction = -1, limits = c(0,78.5))+
  theme_void()+
  guides(fill = guide_colorbar(title.position = "top", title.hjust = 0.5)) +
  labs(title = "Prevalence of Abdominal Obesity in women by state",fill = "Prevalence in % age")

state_womenplot

ggsave(here("Output","map_state_women.svg"), plot = state_womenplot, device = "svg", width = 10, height = 8, units = "in")

# Alternate map: abdominal obesity category
state_womenplot_cat <- ggplot() + 
  geom_sf(data = state_merged_data_with_shape_state_sf, aes(fill = cat_obese_women)) +
  scale_fill_manual(values = color_palette) +
  theme_void() +  
  labs(title = "Prevalence of Abdominal Obesity in Women by state", fill = "Prevalence in % age") 

state_womenplot_cat

ggsave(here("Output","map_state_women_cat.svg"), plot = state_womenplot_cat, device = "svg", width = 10, height = 8, units = "in")
```

```{r session info}
sessioninfo::session_info(pkgs = c("loaded", "attached")[1], to_file = here("Session info","session_info_maps.txt"))
```
